+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- add release function to move changelog unreleased section to new version(pr [#201])
+
 ### Changed
 
 - ci-upgrade jerus-org/circleci-toolkit orb from 0.20.0 to 0.23.0(pr [#200])
@@ -236,6 +240,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#198]: https://github.com/jerus-org/pcu/pull/198
 [#199]: https://github.com/jerus-org/pcu/pull/199
 [#200]: https://github.com/jerus-org/pcu/pull/200
+[#201]: https://github.com/jerus-org/pcu/pull/201
 [Unreleased]: https://github.com/jerus-org/pcu/compare/0.1.10...HEAD
 [0.1.10]: https://github.com/jerus-org/pcu/compare/0.1.9...0.1.10
 [0.1.9]: https://github.com/jerus-org/pcu/compare/0.1.8...0.1.9